feat(docs): update business rules

+++ apis/pool-appointment-api/src/modules/ratings/application/use-cases/delete-rating/delete-rating.use-case.ts
@@ -1,7 +1,7 @@
 import { RatingRepository } from '../../../domain/repositories/rating-repository';
 import { SpaceRepository } from '@/modules/spaces/domain/repositories/space-repository';
 import { UpdateSpaceAverageRating } from '../../../domain/services/update-space-average-rating';
-import { EntityNotFoundError } from '@/core/errors';
+import { EntityNotFoundError, ForbiddenError } from '@/core/errors';
 
 interface DeleteRatingDTO {
   id: string;
@@ -20,11 +20,21 @@ export class DeleteRatingUseCase {
     );
   }
 
-  async execute({ id }: DeleteRatingDTO): Promise<void> {
+  async execute(
+    { id }: DeleteRatingDTO,
+    userRole: string,
+    authenticatedUserId: string,
+  ): Promise<void> {
+    if (userRole !== 'USER') {
+      throw new ForbiddenError('Only users with USER role can delete ratings.');
+    }
     const rating = await this.ratingRepository.findById(id);
 
     if (!rating) {
       throw new EntityNotFoundError('Rating', id);
+    }
+    if (rating.userId !== authenticatedUserId) {
+      throw new ForbiddenError('Users can only delete their own ratings.');
     }
 
     await this.ratingRepository.delete(id);
